app/requirements.txt
flask-pytest==0.0.5
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
